+++ hnix-store-core/src/System/Nix/Hash.hs
@@ -25,12 +25,14 @@ module System.Nix.Hash
   , digestBuilder
   ) where
 
+import Control.DeepSeq (NFData(..))
 import Crypto.Hash (Digest, HashAlgorithm, MD5(..), SHA1(..), SHA256(..), SHA512(..))
 import Data.ByteString (ByteString)
 import Data.Constraint.Extras (Has(has))
 import Data.Constraint.Extras.TH (deriveArgDict)
 import Data.Dependent.Sum (DSum((:=>)))
 import Data.GADT.Compare.TH (deriveGEq, deriveGCompare)
+import Data.GADT.DeepSeq (GNFData(..))
 import Data.GADT.Show.TH (deriveGShow)
 import Data.Kind (Type)
 import Data.Some (Some(Some))
@@ -72,6 +74,16 @@ deriveGEq ''HashAlgo
 deriveGCompare ''HashAlgo
 deriveGShow ''HashAlgo
 deriveArgDict ''HashAlgo
+
+instance NFData (HashAlgo a) where
+  rnf = \case
+    HashAlgo_MD5 -> ()
+    HashAlgo_SHA1 -> ()
+    HashAlgo_SHA256 -> ()
+    HashAlgo_SHA512 -> ()
+
+instance GNFData HashAlgo where
+  grnf = rnf
 
 algoToText :: forall t. HashAlgo t -> Text
 algoToText x = has @NamedAlgo x (algoName @t)